Refactored label frame evener into controller.
Blastn View is very repetitive. I ran into a problem with trying to make
a master width evener. It seems the View object variable needs to be
gridded and the widget must be returned to the same reference or you get
nothing on the screen. Something to investigate for another time.